+++ src/app/api/users/me/rentals/route.ts
@@ -1,0 +1,64 @@
+export async function GET(request: Request) {
+    try {
+        const incomingCookie = request.headers.get("Cookie");
+        const { searchParams } = new URL(request.url);
+        const role = searchParams.get("role") || "";
+
+        const response = await fetch(
+            process.env.NEXT_PUBLIC_API_URL +
+                `/api/users/me/rentals?role=${encodeURIComponent(role)}`,
+            {
+                method: "GET",
+                credentials: "include",
+                headers: {
+                    "Content-Type": "application/json",
+                    Cookie: incomingCookie || "",
+                },
+            },
+        );
+
+        let responseData: unknown = null;
+        try {
+            responseData = await response.json();
+        } catch {
+            responseData = null;
+        }
+
+        if (!response.ok) {
+            const errorMessage =
+                typeof responseData === "object" &&
+                responseData !== null &&
+                "message" in responseData &&
+                typeof (responseData as { message: unknown }).message ===
+                    "string"
+                    ? (responseData as { message: string }).message
+                    : "대여 내역 조회 API 호출에 실패했습니다.";
+
+            return Response.json(
+                {
+                    success: false,
+                    code: "RENTAL_FETCH_FAILED",
+                    message: errorMessage,
+                    data: null,
+                },
+                { status: response.status },
+            );
+        }
+
+        return Response.json(responseData, { status: response.status });
+    } catch (error) {
+        console.error("Error in GET /api/users/me/rentals:", error);
+        return Response.json(
+            {
+                success: false,
+                code: "INTERNAL_SERVER_ERROR",
+                message:
+                    error instanceof Error
+                        ? error.message
+                        : "Internal server error",
+                data: null,
+            },
+            { status: 500 },
+        );
+    }
+}
