+++ src/app/items/rent/history/_component/HydrateClient.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { HydrationBoundary } from "@tanstack/react-query";
+import { Rental } from "@/types/common";
+import RentalHistoryClient from "./RentalHistoryClient";
+
+export default function HydrateClient({ state }: { state: any }) {
+    return (
+        <HydrationBoundary state={state}>
+            <RentalHistoryClient />
+        </HydrationBoundary>
+    );
+}
